pyproject.toml
[tool.poetry]
name = "university-chatbot"
version = "0.1.0"
description = "LangChain RAG chatbot for university PDFs with API and UI"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "chatbot" }]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.1.14"
langgraph = "^0.0.35"
openai = "^1.23.4"
chromadb = "^0.4.24"
python-dotenv = "^1.0.1"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
groq = "^0.24.0"
pypdf = "^5.4.0"
sentence-transformers = "^4.1.0"
python-multipart="^0.0.20"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"